+++ src/App.js
@@ -1,97 +1,9 @@
-import Square from "./components/Square";
-import { useState } from "react";
+import Game from "./components/Game"
 
-function Board({ xIsNext, squares, onPlay }) {
-  const handleClick = (i) => {
-    if (squares[i] || calculateWinner(squares)) return;
+export default function App(){
 
-    const nextSquares = squares.slice();
-    xIsNext ? (nextSquares[i] = "X") : (nextSquares[i] = "O");
-
-    onPlay(nextSquares);
-  };
-
-  const winner = calculateWinner(squares);
-  let status;
-  if (winner) {
-    status = "Winner: " + winner;
-  } else {
-    status = "Next player: " + (xIsNext ? "X" : "O");
-  }
-
-  return (
-    <>
-      <h1 className="status">{status}</h1>
-      <div className="board-row">
-        <Square value={squares[0]} onSquareClick={() => handleClick(0)} />
-        <Square value={squares[1]} onSquareClick={() => handleClick(1)} />
-        <Square value={squares[2]} onSquareClick={() => handleClick(2)} />
-      </div>
-      <div className="board-row">
-        <Square value={squares[3]} onSquareClick={() => handleClick(3)} />
-        <Square value={squares[4]} onSquareClick={() => handleClick(4)} />
-        <Square value={squares[5]} onSquareClick={() => handleClick(5)} />
-      </div>
-      <div className="board-row">
-        <Square value={squares[6]} onSquareClick={() => handleClick(6)} />
-        <Square value={squares[7]} onSquareClick={() => handleClick(7)} />
-        <Square value={squares[8]} onSquareClick={() => handleClick(8)} />
-      </div>
-    </>
-  );
+  return Game()
 }
 
-export default function Game() {
-  const [history, setHistory] = useState([Array(9).fill(null)]);
-  const [currentMove, setCurrentMove] = useState(0) 
-  const currentSquares = history[currentMove];
-  const xIsNext = currentMove % 2 === 0;
 
-  function handlePlay(nextSquares) {
-    const nextHistory = [...history.slice(0, currentMove + 1), nextSquares];
-    setHistory(nextHistory)
-    setCurrentMove(nextHistory.length - 1)
-  }
 
-  const getButtonDescription = (move) =>
-    move > 0 ? `Go to move #${move}` : "Go to game start";
-
-  const moves = history.map((squares, move) => (
-    <li key={move}>
-      <button className="travel-in-time-button" onClick={() => setCurrentMove(move)}>{getButtonDescription(move)}</button>
-    </li>
-  ));
-
-  return (
-    <div className="game">
-      <div className="game-board">
-        <Board xIsNext={xIsNext} squares={currentSquares} onPlay={handlePlay} />
-        <div className="current-move">{currentMove === 0 ? "You are at the start move" : `You are at move #${currentMove}`}</div>
-      </div>
-
-      <div className="game-info">
-        <ol className="travel-in-time-button-list">{moves}</ol>
-      </div>
-    </div>
-  );
-}
-
-const calculateWinner = (squares) => {
-  const lines = [
-    [0, 1, 2],
-    [3, 4, 5],
-    [6, 7, 8],
-    [0, 3, 6],
-    [1, 4, 7],
-    [2, 5, 8],
-    [0, 4, 8],
-    [2, 4, 6],
-  ];
-  for (let i = 0; i < lines.length; i++) {
-    const [a, b, c] = lines[i];
-    if (squares[a] && squares[a] === squares[b] && squares[a] === squares[c]) {
-      return squares[a];
-    }
-  }
-  return null;
-};
