+++ src/helpers/Constants.js
@@ -1,0 +1,2 @@
+export const NUMBER_OF_COLS = 3;
+export const NUMBER_OF_ROWS = 3;
