+++ src/styles.css
@@ -97,43 +97,3 @@ ul {
   border: 1px solid gray;
   padding: 2px 4px;
 }
-
-button {
-  /* Remove background */
-  background: none;
-  background-color: transparent; /* Ensure full transparency */
-
-  /* Remove border */
-  border: none;
-
-  /* Remove default padding and margin */
-  padding: 0;
-  margin: 0;
-
-  /* Reset font properties to inherit from parent */
-  font: inherit;
-  color: inherit;
-  text-align: inherit; /* Useful for centering text in a button */
-
-  /* Remove default cursor (optional, often you want a pointer) */
-  /* cursor: default; */ /* or cursor: auto; */
-
-  /* Remove outline on focus (be careful with accessibility!) */
-  /* For accessibility, it's generally better to style the outline
-     rather than removing it completely. */
-  outline: none; 
-
-  /* Remove box shadow */
-  box-shadow: none;
-
-  /* Remove default text decoration (e.g., underline on some links/buttons) */
-  text-decoration: none;
-
-  /* Reset appearance property (specific to some browsers) */
-  -webkit-appearance: none; /* For WebKit browsers (Chrome, Safari) */
-  -moz-appearance: none;    /* For Firefox */
-  appearance: none;         /* Standard property */
-
-  /* Set line-height to ensure consistent vertical alignment (optional but good practice) */
-  line-height: 1.15; 
-}
